Fixed case when rendering an array with no elements
In the old code, nothing got rendered if the template was asked to
render an array with no elements. This loses the block, and retaining
the block is essential to correct operation of the template. The new
code handles this as a special case and writes the surrounding block
tags, unless this is a with block.