Modify request so list of page links can be properly built
To indicate that update_records should delete a record, the oldid
field should be set and the id field unset. Prior to this change,
these were not done. Also, the unit test fro remove in DirData.t did
not pass the request. This was a case where the test was not changed
to match a change in the interface.

and the id field should be unset.