Got PostData.pm to pass unit tests
When coding PostData.pm, I discovered that the object it was
subclassing (PageData) was not flexible enough to subclass
easily. Which caused me to go back and rewrite the methods used when
setting up the data to render. After getting the code to work, I
rewrote PostData to use the new methods. I aslo wrote BlogData.pm,
which subclasses DirData, to handle the browse and search methods for
posts.