Write complete id into link
The code uses id field to track links. It now uses the complete link, so that
there is no ambiguity int the parent links. There was also a problem handling
oldid, which indicates a record should be deleted, which was fixed. And I added
id to the record returned by read_primary, so that links can be generated from
it.