A better fix for the broken link problem
The cause of the problem was the difference between how mask_template
and render decide to ignore sections in the template. Mask_template
was ignoring sections if they did not exist in the data, render was
ignoring them if the data was undefined. This allowed the template to
show through in the rendered output where the data was equal to undef,
a possible result from build_records in PageData. The fix has
mask_template use the same test as render and restores the previous
behavior (returning undef when build_records does not change the list
of records.)