A fix for the pluralization problem in command links
The fix require changeing has_one_subtype so that it returns the one
type when it is true, undef otherwise. This is responsible for most of
the changes. The code now creates the subobject when we have a parent
command and there is only one type so it can read the plural parameter
from it.