Fix redirect bug after remove
The code which removes a page also deletes the id from the object. As
a result, the redirect used to go to the top page. The new code saves
the id before calling remove_data. The same change was made to
AddCommand and RemoveCommand, even though AddCommand doesn't need the
fix.